Fix "View more" link on the Jobs/Show page
